\name{getRunorder}
\alias{getRunorder}
\title{
Calculate the order in which functions need to be run
}
\description{
Order of functions depends on
}
\usage{
function (fun.calls, current.state, pars, return.output = F)}
\arguments{
\item{fun.calls}{output of parseCalls}
\item{current.state}{environment with current state variables}
\item{pars}{pars}
\item{return.output}{whether to return the results from the function runs required to establish the run order}
}
\details{
Placeholder help file
}
\value{
character vector of function names in the order in which they can be run
}
\references{
}
\author{
Joseph Guillaume
}
\seealso{
Summary of \code{\link{workflow}} tools
}
\examples{
## Not run.
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
