\name{jgutils-package}
\alias{jgutils-package}
\alias{jgutils}
\docType{package}
\title{
Collection of utilities
}
\description{
Utilities by Joseph Guillaume
}
\details{
\tabular{ll}{
Package: \tab jgutils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-30\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

Mapping functions
\code{\link{mapping}}

Workflow functions
\code{\link{workflow}}

Timeseries functions
\code{\link{timeseries}}

BOM data & Environment functions

Uncertainty functions
\code{\link{pavingBounds}}
}
\author{
Joseph Guillaume
}
