\name{textRect}
\alias{textRect}
\alias{lattice}
\title{
Textbox with a background colour
}
\description{
Panel function to create a textbox with a background colour
}
\usage{
textRect(x, y, ss, lty = 0, pos = NULL, adj = c(0.5, 0.5), offset = 0.5, col="white",
    cex = add.text$cex)
}
\arguments{
  \item{x}{
x coordinate
}
  \item{edges}{
y coordinate
}
  \item{ss}{
String to use for labels
}
}
\author{
Joseph Guillaume
}

